int64 indexing: API methods and bug fixes
- np.eye: add long overload for N/M params, int delegates to long
- np.identity: add long overload delegating to np.eye(long)
- np.are_broadcastable: add long[] overload for shapes
- NDArray.matrix_power: remove (int) cast to shape[0] since np.eye accepts long
- DefaultEngine.ReductionOp: fix argmax/argmin unboxing from (long)(int) to (long)
  - argmax_elementwise/argmin_elementwise return long, not int
  - Previous code tried to unbox long as int, causing InvalidCastException
- np.isinf tests: mark with [OpenBugs] since np.isinf is dead code

+++ src/NumSharp.Core/LinearAlgebra/NDArray.matrix_power.cs
@@ -14,7 +14,7 @@ namespace NumSharp
             for (int idx = 2; idx <= power; idx++)
                 product = TensorEngine.Dot(product, this);
 
-            product = (power == 0) ? np.eye((int)product.shape[0]) : product;
+            product = (power == 0) ? np.eye(product.shape[0]) : product;
 
             return product;
         }
